Return test result record from run-test-case.
Return a map describing the outcome of running the test case instead
of writing directly to the console. Move test output into new
display-test-result function. Calculate a summary containing the total
number of passed, failed and ignored tests which is displayed at the
end of test execution. Use the number of failed tests as the return
code for the process.